backend/pyproject.toml
[project]
name = "fnol-backend"
version = "0.1.0"
description = "FNOL AI Agents — FastAPI + LangGraph backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "copilotkit>=0.1.78",
    "langgraph>=0.3.25",
    "langchain>=0.3.14",
    "langchain-openai>=0.3.0",
    "openai>=1.58.0",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.0",
    "pytesseract>=0.3.13",
    "pillow>=11.1.0",
    "PyPDF2>=3.0.1",
    "httpx>=0.28.1",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.1",
]
